+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.5] Q4 2023
+- copy-from-image: fixed integration test
+
 ## [0.16.4] Q4 2023
 - ssh tunnel:
   - set 'https://cp.omnect.conplement.cloud' as default backend